+++ src/components/VerseEditorDialog/EditScreen.js
